+++ ChainOfInvocatios/Pages/DashboardPage.cs
@@ -1,0 +1,40 @@
+using OpenQA.Selenium;
+
+namespace ChainOfInvocatios.Pages
+{
+    public class DashboardPage : BasePage
+    {
+        private static string END_POINT = "index.php?/dashboard";
+        
+        // Описание элементов
+        private static readonly By SidebarProjectsAddButtonBy = By.Id("sidebar-projects-add");
+        
+        public DashboardPage(IWebDriver driver, bool openPageByUrl) : base(driver, openPageByUrl)
+        {
+        }
+
+        public DashboardPage(IWebDriver driver) : base(driver)
+        {
+        }
+
+        public override bool IsPageOpened()
+        {
+            try
+            {
+                return SidebarProjectsAddButton.Displayed;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+            
+        }
+
+        protected override string GetEndpoint()
+        {
+            return END_POINT;
+        }
+
+        public IWebElement SidebarProjectsAddButton => WaitsHelper.WaitForExists(SidebarProjectsAddButtonBy);
+    }
+}
