+++ ChainOfInvocatios/Pages/ProjectPages/ProjectBasePage.cs
@@ -1,0 +1,26 @@
+using OpenQA.Selenium;
+
+namespace ChainOfInvocatios.Pages.ProjectPages;
+
+public class ProjectBasePage : BasePage
+{
+    private static readonly By NameInputBy = By.Id("name");
+    
+    public ProjectBasePage(IWebDriver driver) : base(driver)
+    {
+    }
+
+    public ProjectBasePage(IWebDriver driver, bool openPageByUrl) : base(driver, openPageByUrl)
+    {
+    }
+
+    protected override string GetEndpoint()
+    {
+        throw new NotImplementedException();
+    }
+
+    public override bool IsPageOpened()
+    {
+        throw new NotImplementedException();
+    }
+}
