+++ ChainOfInvocatios/Steps/BaseStep.cs
@@ -1,0 +1,13 @@
+using ChainOfInvocatios.Pages;
+using ChainOfInvocatios.Pages.ProjectPages;
+using OpenQA.Selenium;
+
+namespace ChainOfInvocatios.Steps;
+
+public class BaseStep(IWebDriver driver)
+{
+    protected readonly IWebDriver Driver = driver;
+   protected LoginPage? LoginPage { get; set; }
+   protected DashboardPage? DashboardPage { get; set; }
+   protected AddProjectPage? AddProjectPage { get; set; }
+}
