+++ ChainOfInvocatios/Tests/BaseTest.cs
@@ -1,0 +1,35 @@
+using ChainOfInvocatios.Core;
+using ChainOfInvocatios.Helpers.Configuration;
+using ChainOfInvocatios.Steps;
+using OpenQA.Selenium;
+using ChainOfInvocatios.Helpers;
+
+namespace ChainOfInvocatios.Tests;
+
+[Parallelizable(scope: ParallelScope.All)]
+[FixtureLifeCycle(LifeCycle.InstancePerTestCase)]
+public class BaseTest
+{
+    protected IWebDriver Driver { get; private set; }
+
+    protected NavigationSteps _navigationSteps;
+    protected ProjectSteps _projectSteps;
+
+    [SetUp]
+    public void Setup()
+    {
+        Driver = new Browser().Driver;
+
+        _navigationSteps = new NavigationSteps(Driver);
+        _projectSteps = new ProjectSteps(Driver);
+        
+        
+        Driver.Navigate().GoToUrl(Configurator.AppSettings.URL);
+    }
+
+    [TearDown]
+    public void TearDown()
+    {
+        Driver.Quit();
+    }
+}
