+++ ChainOfInvocatios/Tests/LoginTest.cs
@@ -1,0 +1,13 @@
+using ChainOfInvocatios.Helpers.Configuration;
+
+namespace ChainOfInvocatios.Tests;
+
+public class LoginTest : BaseTest
+{
+    [Test]
+    public void SuccessLoginTest()
+    {
+        _navigationSteps
+            .SuccessfulLogin(Configurator.AppSettings.Username, Configurator.AppSettings.Password);
+    }
+}
