+++ LoadableComponent/Helpers/Configuration/AppSettings.cs
@@ -1,0 +1,8 @@
+namespace PageFactory.Helpers.Configuration;
+
+public class AppSettings
+{
+    public string? URL { get; set; }
+    public string? Username { get; set; }
+    public string? Password { get; set; }
+}
