+++ LoadableComponent/UnitTest1.cs
@@ -1,0 +1,15 @@
+namespace LoadableComponent;
+
+public class Tests
+{
+    [SetUp]
+    public void Setup()
+    {
+    }
+
+    [Test]
+    public void Test1()
+    {
+        Assert.Pass();
+    }
+}
