+++ LoadableComponent/appsettings.json
@@ -1,0 +1,16 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
+    }
+  },
+  "AppSettings": {
+    "URL": "https://aqa2503.testrail.io/",
+    "Username": "atrostyanko@gmail.com",
+    "Password": "Americana#1978"
+  },
+  "BrowserType": "chrome",
+  "WaitsTimeout": 10
+}
