--- PageObjectSimple/Core/Browser.cs
@@ -1,24 +1,0 @@
-﻿using OpenQA.Selenium;
-using PageObjectSimple.Helpers.Configuration;
-
-namespace PageObjectSimple.Core
-{
-    public class Browser
-    {
-        public IWebDriver Driver { get; }
-
-        public Browser()
-        {
-            Driver = Configurator.BrowserType?.ToLower() switch
-            {
-                "chrome" => new DriverFactory().GetChromeDriver(),
-                "firefox" => new DriverFactory().GetFirefoxDriver(),
-                _ => Driver
-            } ?? throw new InvalidOperationException("Browser is not supported.");
-
-            Driver.Manage().Window.Maximize();
-            Driver.Manage().Cookies.DeleteAllCookies();
-            //Driver!.Manage().Timeouts().ImplicitWait = TimeSpan.FromSeconds(0);
-        }
-    }
-}
