--- PageObjectSimple/Helpers/Configuration/AppSettings.cs
@@ -1,8 +1,0 @@
-namespace PageObjectSimple.Helpers.Configuration;
-
-public class AppSettings
-{
-    public string? URL { get; set; }
-    public string? Username { get; set; }
-    public string? Password { get; set; }
-}
