--- PageObjectSimple/Steps/BaseStep.cs
@@ -1,19 +1,0 @@
-using OpenQA.Selenium;
-using PageObjectSimple.Pages;
-using PageObjectSimple.Pages.ProjectPages;
-
-namespace PageObjectSimple.Steps;
-
-public class BaseStep
-{
-    protected IWebDriver Driver;
-
-    public LoginPage LoginPage => new LoginPage(Driver);
-    public DashboardPage DashboardPage => new DashboardPage(Driver);
-    public AddProjectPage AddProjectPage => new AddProjectPage(Driver);
-
-    public BaseStep(IWebDriver driver)
-    {
-        Driver = driver;
-    }
-}
