--- PageObjectSimple/Steps/ProjectSteps.cs
@@ -1,16 +1,0 @@
-using OpenQA.Selenium;
-using PageObjectSimple.Pages.ProjectPages;
-
-namespace PageObjectSimple.Steps;
-
-public class ProjectSteps : BaseStep
-{
-    public ProjectSteps(IWebDriver driver) : base(driver)
-    {
-    }
-
-    public void NavigateToAddProjectPage()
-    {
-        new AddProjectPage(Driver, true);
-    }
-}
