--- PageObjectSimple/Tests/ProjectsTest.cs
@@ -1,15 +1,0 @@
-using PageObjectSimple.Helpers.Configuration;
-
-namespace PageObjectSimple.Tests;
-
-public class ProjectsTest : BaseTest
-{
-    [Test]
-    public void CreateProjectTest()
-    {
-        NavigationSteps.NavigateToLoginPage();
-        NavigationSteps.SuccessfulLogin(Configurator.AppSettings.Username, Configurator.AppSettings.Password);
-        
-        ProjectSteps.NavigateToAddProjectPage();
-    }
-}
