+++ TMS_AQA.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.7.34221.43
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PageObjectSimple", "PageObjectSimple\PageObjectSimple.csproj", "{7A349357-5F50-4753-B9A9-1A7FE7F964F9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChainOfInvocatios", "ChainOfInvocatios\ChainOfInvocatios.csproj", "{7A349357-5F50-4753-B9A9-1A7FE7F964F9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LoadableComponent", "LoadableComponent\LoadableComponent.csproj", "{019B1080-6655-473B-8E1B-38F1464EC26A}"
 EndProject
 Global
 	GlobalSection(SolutionProperties) = preSolution
@@ -33,5 +35,17 @@ Global
 		{7A349357-5F50-4753-B9A9-1A7FE7F964F9}.Release|x64.Build.0 = Release|Any CPU
 		{7A349357-5F50-4753-B9A9-1A7FE7F964F9}.Release|x86.ActiveCfg = Release|Any CPU
 		{7A349357-5F50-4753-B9A9-1A7FE7F964F9}.Release|x86.Build.0 = Release|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Debug|x64.Build.0 = Debug|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Debug|x86.Build.0 = Debug|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Release|x64.ActiveCfg = Release|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Release|x64.Build.0 = Release|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Release|x86.ActiveCfg = Release|Any CPU
+		{019B1080-6655-473B-8E1B-38F1464EC26A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
